+++ src/Bridges/Constraint/NonOverlappingOrthotopes/noov_to_noovcond.jl
@@ -1,18 +1,18 @@
 """
-Bridges `CP.NonOverlappingOrthotopes` to 
-`CP.ConditionallyNonOverlappingOrthotopes`.
+Bridges `CP.NonOverlappingOrthotopes{CP.UNCONDITIONAL_NONVERLAPPING_ORTHOTOPES}` to 
+`CP.NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES}`.
 """
 struct NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T} <: MOIBC.AbstractBridge
     var::MOI.VariableIndex
     var_con::MOI.ConstraintIndex{MOI.SingleVariable, MOI.EqualTo{T}}
-    con::MOI.ConstraintIndex{MOI.VectorAffineFunction{T}, CP.ConditionallyNonOverlappingOrthotopes}
+    con::MOI.ConstraintIndex{MOI.VectorAffineFunction{T}, CP.NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES}}
 end
 
 function MOIBC.bridge_constraint(
     ::Type{NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T}},
     model,
     f::MOI.VectorOfVariables,
-    s::CP.NonOverlappingOrthotopes
+    s::CP.NonOverlappingOrthotopes{CP.UNCONDITIONAL_NONVERLAPPING_ORTHOTOPES}
 ) where {T}
     return MOIBC.bridge_constraint(
         NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T},
@@ -26,15 +26,16 @@ function MOIBC.bridge_constraint(
     ::Type{NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T}},
     model,
     f::MOI.VectorAffineFunction{T},
-    s::CP.NonOverlappingOrthotopes
+    s::CP.NonOverlappingOrthotopes{CP.UNCONDITIONAL_NONVERLAPPING_ORTHOTOPES}
 ) where {T}
     f_scalars = MOIU.scalarize(f)
 
     # One new variable that takes the value "true".
     var, var_con = MOI.add_constrained_variable(model, MOI.EqualTo(one(T)))
 
-    # The equivalent ConditionallyNonOverlappingOrthotopes with the conditions
-    # set to true.
+    # The equivalent 
+    # NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES} with 
+    # the conditions set to true.
     new_f = MOI.ScalarAffineFunction{T}[]
     j = 1
     for _ in 1:s.n_orthotopes
@@ -49,7 +50,7 @@ function MOIBC.bridge_constraint(
     con = MOI.add_constraint(
         model, 
         new_f,
-        CP.ConditionallyNonOverlappingOrthotopes(s.n_orthotopes, s.n_dimensions)
+        CP.NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES}(s.n_orthotopes, s.n_dimensions)
     )
 
     return NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge(var, var_con, con)
@@ -58,7 +59,7 @@ end
 function MOI.supports_constraint(
     ::Type{NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T}},
     ::Union{Type{MOI.VectorOfVariables}, Type{MOI.VectorAffineFunction{T}}},
-    ::Type{CP.NonOverlappingOrthotopes},
+    ::Type{CP.NonOverlappingOrthotopes{CP.UNCONDITIONAL_NONVERLAPPING_ORTHOTOPES}},
 ) where {T}
     return true
 end
@@ -69,7 +70,7 @@ end
 
 function MOIB.added_constraint_types(::Type{NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T}}) where {T}
     return [
-        (MOI.VectorAffineFunction{T}, CP.ConditionallyNonOverlappingOrthotopes),
+        (MOI.VectorAffineFunction{T}, CP.NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES}),
     ]
 end
 
@@ -89,7 +90,7 @@ end
 function MOI.get(
     ::NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T},
     ::MOI.NumberOfConstraints{
-        MOI.VectorAffineFunction{T}, CP.ConditionallyNonOverlappingOrthotopes,
+        MOI.VectorAffineFunction{T}, CP.NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES},
     },
 ) where {T}
     return 1
@@ -114,7 +115,7 @@ end
 function MOI.get(
     b::NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge{T},
     ::MOI.ListOfConstraintIndices{
-        MOI.VectorAffineFunction{T}, CP.ConditionallyNonOverlappingOrthotopes,
+        MOI.VectorAffineFunction{T}, CP.NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES},
     },
 ) where {T}
     return [b.con]